Extract functionality to separate files
